# feat: Support optimised init from non-dict `Mapping` objects in `from_records` and frame/series constructors