--- a/py-polars/polars/_utils/construction/utils.py
+++ b/py-polars/polars/_utils/construction/utils.py
@@ -1,13 +1,12 @@
 from __future__ import annotations
 
+from collections.abc import Sequence
 from functools import lru_cache
 from typing import TYPE_CHECKING, Any, Callable, get_type_hints
 
 from polars.dependencies import _check_for_pydantic, pydantic
 
 if TYPE_CHECKING:
-    from collections.abc import Sequence
-
     import pandas as pd
 
 PANDAS_SIMPLE_NUMPY_DTYPES = {
@@ -48,7 +47,7 @@ def try_get_type_hints(obj: type) -> dict[str, Any]:
 
 @lru_cache(64)
 def is_namedtuple(cls: Any, *, annotated: bool = False) -> bool:
-    """Check whether given class derives from NamedTuple."""
+    """Check if given class derives from NamedTuple."""
     if all(hasattr(cls, attr) for attr in ("_fields", "_field_defaults", "_replace")):
         if not isinstance(cls._fields, property):
             if not annotated or len(cls.__annotations__) == len(cls._fields):
@@ -57,13 +56,15 @@ def is_namedtuple(cls: Any, *, annotated: bool = False) -> bool:
 
 
 def is_pydantic_model(value: Any) -> bool:
-    """Check whether value derives from a pydantic.BaseModel."""
+    """Check if value derives from a pydantic.BaseModel."""
     return _check_for_pydantic(value) and isinstance(value, pydantic.BaseModel)
 
 
-def is_sqlalchemy(value: Any) -> bool:
-    """Check whether value is an instance of a SQLAlchemy object."""
-    return getattr(value, "__module__", "").startswith("sqlalchemy.")
+def is_sqlalchemy_row(value: Any) -> bool:
+    """Check if value is an instance of a SQLAlchemy sequence or mapping object."""
+    return getattr(value, "__module__", "").startswith("sqlalchemy.") and isinstance(
+        value, Sequence
+    )
 
 
 def get_first_non_none(values: Sequence[Any | None]) -> Any: