@alexander-beedie 2025-05-07
Looks like it needs a small update to handle `None` entries; will add that shortly.